Move more code, create needed stuff
