+++ Assets/deCrawl/Scrips/Utils/InterfaceFinder.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cbd4dc0ab6f4187469b95d9e0802c32e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
