+++ Assets/deCrawl/Scrips/World/IMovingEntity.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0adf1dc9146c9cd46bd44c5ab31357b2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
