+++ Assets/deCrawl/Scrips/World/NavInstructions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9ecb0277af1c8d94da9fd57b5106126d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
